fix: proper range handling for :Fyler command
Previously, we were *always* using the last visual selection as the directory,
even if the Fyler command not called from a visual mode mapping, and
also ignored the range passed to the command (e.g. `:2,3Fyler`).

This improves the behavior by:

- Checking if the command was called from a visual mode mapping, and if
  so, using the visual selection (including column offsets) as the
  directory.
- Otherwise, checking if the passed range matches the last visual
  selection, and if so, using the visual selection (including column
  offsets) as the directory. This is to support the case of `:'<,'>Fyler`
- Otherwise, if a line-wise range was passed, using the lines as the
  directory.

In all of the above cases, if the range includes multiple lines, we only
use the first line of the range. While directory paths can sometimes
contain newlines, this is exceedingly rare. In the future, we may want to
treat the full range as multiple directories to open, but it's not clear how
opening multiple directories would work currently.

Fixes #265